A ton of API cleanups
When we started this we were pretty new to rust and just clone()
everywhere. I went through and found everywhere we should be able to
references and fixed that. Should make us slightly more effienct and
make the APIs a little cleaner to user (less clone()s everywhere).

Sorry for the mondo PR