Remove RGS workarounds
We don't need to call `remove_stale_channels_and_tracking_with_time` anymore because syncing RGS will now do that as of ldk 0.0.115.

We also do not need to check if the last sync was 24 hours ago because the RGS server will send us an empty diff if there is nothing to sync.