Encode LDK channel manager + monitor as hex
This is a more mimimal encoding for the two data types so it should take
up less space in storage, maybe will cause less storage issues.

I added it so if it sees an array of numbers it'll handle it correctly
so we don't have backwards compatiblity issues.

Not sure if this will effect force close stuff but at the very least,
it'll make storage more effienct.