Handle payjoin errors according to BIP 78
"The receiver is allowed to return implementation specific errors which
may assist the sender to diagnose any issue.

However, it is important that error codes that are not well-known and
that the message do not appear on the sender's software user interface.
Such error codes or messages could be used maliciously to phish a non-
technical user. Instead those errors or messages can only appear in
debug logs.

It is advised to hard code the description of the well known error
codes into the sender's software."

See:
https://github.com/bitcoin/bips/blob/master/bip-0078.mediawiki#receivers-well-known-errors

This commit displays templates based on a particular \`ResponseError\`
and debugs the rest, according to this specification.